backend/requirements.txt
djangorestframework==3.12.4
djoser==2.1.0
python-dotenv==0.19.0
gunicorn==20.1.0
